deletes only the resume that belongs to the user

+++ src/controllers/cv.controller.ts
@@ -82,7 +82,7 @@ const cvController = {
         const userResume = await Resume.findOne({user});
         if((userResume?._id)?.toString() !== id){
             return res.status(401).json({
-                created: false,
+                updated: false,
                 errors: [
                     {
                         value: user,
@@ -104,8 +104,23 @@ const cvController = {
             result: resume
         });
     },
-    delete: async (req: Request, res: Response) => {
+    delete: async (req: CustomRequest, res: Response) => {
         const {id} = req.params;
+        const user = req.user;
+        const userResume = await Resume.findOne({user});
+        if((userResume?._id)?.toString() !== id){
+            return res.status(401).json({
+                deleted: false,
+                errors: [
+                    {
+                        value: user,
+                        msg: 'No puede eliminar un currículo perteneciente a otro usuario.',
+                        param: 'user',
+                        location: 'body'
+                    }
+                ]
+            });
+        }
         const resume = await Resume.findByIdAndDelete(id);
         return res.status(200).json({
             deleted: true,
